src/requirements-dev.txt
-r requirements.txt
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0  # For HTTP tests